Refactor tests to mirror modular src/tools/ structure
Split monolithic test_main.py (815 lines, 34 classes) into 8 focused
test files under tests/tools/, one per src/tools/ module. Update
test_api_client.py imports to use canonical paths. Remove dead re-export
block from src/main.py. Fixes accidental standalone test collection
(114 → 113 tests).
